// resources/js/types/profile.ts
export type Profile = {
    id?: number;
    name?: any;
    kode_faskes?: any;
    email?: any;
    phone_number?: any;
    address?: any;
    province_id?: any;
    province_name?: any;
    city_id?: any;
    city_name?: any;
    district_id?: any;
    district_name?: any;
    village_id?: any;
    village_name?: any;
    created_at?: Date;
    updated_at?: Date;
    deleted_at?: null;
}